// src/app/models/offer-item.model.ts
export interface OfferItems {
    modes: string[];
    movementTypes: string[];
    incoterms: string[];
    countriesCities: {
        USA: string[];
        China: string[];
        Turkey: string[];
    };
    packageTypes: string[];
    unit1: string[];
    unit2: string[];
    currencies: string[];
}