// src/app/models/offer-list.model.ts
export interface OfferListItems {
    id: string;
    mode: string;
    movementType: string;
    incoterms: string;
    countriesCities: string;
    packageType: string;
    unit1: string;
    unit1Value: string;
    unit2: string;
    unit2Value: string;
    currency: string;
    palletCount: number;
}